rename modules to be lowercase
